move hydrator tests into correct crate
